daq_core/requirements.txt
# DAQ Core 依赖
paho-mqtt>=2.0.0
